requirements.txt
streamlit==1.39.0
tensorflow==2.17.0
pillow==10.3.0
numpy==1.26.4
pandas==2.2.2
